api/requirements.txt
attrs==17.3.0
beautifulsoup4==4.6.0
coverage==4.4.2
hupper==1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.4.2
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pyathena==1.2.0
Pygments==2.2.0
pyramid==1.9.1
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pytest==3.3.1
pytest-cov==2.5.1
repoze.lru==0.7
six==1.11.0
SQLAlchemy==1.1.6
translationstring==1.3
venusian==1.1.0
WebOb==1.7.4
WebTest==2.0.29
wsgicors==0.7.0
zope.deprecation==4.3.0
zope.interface==4.4.3
gns3-netifaces==0.10.4.1
pycryptodome==3.4.5
PyJWT==1.4.2
uwsgi==2.0.15